requirements.txt
autopep8==2.0.1
openpyxl==3.0.10
pandas==1.5.2
pylint==2.15.8
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
